chore: remove deprecated Hibernate enhancement option and redundant long casts
Category 7 - enableAssociationManagement deprecated (Hibernate 6.4+):
Remove enableAssociationManagement = true from 5 build.gradle files.
Codebase uses records + manual mappers so bidirectional association
management bytecode enhancement has no effect. Defaults to false.

Category 8 - redundant (long) cast:
- CalPremiumManagerTest.java:358 ids[i] = (long)(i+1) → i+1  (int→long widening, n=10 no overflow)
- TickBufferTest.java:116 tick((long)(wId*100+i)) → tick(wId*100+i)  (max=419 no overflow)

+++ modules/market_data/src/test/java/com/example/demo/market_data/application/usecase/CalPremiumManagerTest.java
@@ -355,7 +355,7 @@ class CalPremiumManagerTest {
         int n = 10;
         long[] ids = new long[n];
         for (int i = 0; i < n; i++) {
-            ids[i] = (long) (i + 1);
+            ids[i] = i + 1;
             final long id = ids[i];
             given(codeGetter.get(id)).willReturn(Optional.of(mc(99L, "ETH")));
             given(tickGetter.get(id)).willReturn(Optional.of(tick(id, "2000", "2001")));
